refactor: determine_wall_direction function split into multiple functions
